# Optimized event_based_risk for covs != 0