Continuation of https://github.com/gem/oq-engine/pull/6607. Here are the figures for Miriam's island with no correlation:
```
# master
calc_13579, maxmem=4.7 GB    time_sec memory_mb counts
============================ ======== ========= ======
total event_based_risk       534      88        40    
computing risk               504      0.0       240   
EventBasedRiskCalculator.run 67       86        1     
aggregating losses           12       0.0       240   
averaging losses             6.39794  0.0       240   

# this branch
calc_13578, maxmem=3.4 GB    time_sec memory_mb counts
============================ ======== ========= ======
total event_based_risk       84       117       40    
computing risk               53       0.0       240   
aggregating losses           14       0.0       240   
EventBasedRiskCalculator.run 12       86        1     
averaging losses             5.11174  0.0       240   
```
The speedup in "computing risk" is 9.5x. We are also using a lot less memory. We have now finally beaten engine-3.11:
```
# engine 3.11
calc_13581, maxmem=3.0 GB    time_sec memory_mb counts
============================ ======== ========= ======
total event_based_risk       76       97        21    
computing risk               76       0.0       512   
EventBasedRiskCalculator.run 27       405       1      
aggregating losses           1.95674  0.0       21   
```
(tthe total runtimes is 12 s vs 27s having saved a lot of time not generating the epsilon matrix). There is still room for improvement in "aggregating losses".